fix(lint): move global declare block to d.ts file
