+++ backend/express.d.ts
@@ -1,0 +1,11 @@
+declare global {
+    namespace Express {
+        interface Request {
+            user?: string;
+            admin?: boolean;
+            email?: string;
+            dayID?: number;
+        }
+    }
+}
+export {};
